requirements.txt
torch
diffusers
scikit-learn
pandas
numpy
antspyx
nibabel
nilearn
atexit
signal